readme suggestions
This project is off to a good start, but does not demonstrate that you used many of the tools we learned in class, even though there were clear opportunities in the project to do so.  A stronger project would have: created a function, used apply, used a loop, used bash or automated activities in some respect.